Fixed A Couple of Things.
Line 8: The line should say "Don't use dynamic terms", the dynamic terms do not help the devs in the slightest, because if the user is on an outdated "latest" version, the devs may waste valuable time trying to figure out why it is still "happening" when it isn't.

Line 29: The line shouldn't tell users to ignore the build number, only the device and operating system, because if a user is ignoring the build number, and the user who reported it is not on the latest build, then the devs might have fixed it already but the user hasn't updated yet. It got a bit longer, but we shouldn't be worrying about length over detail.